requirements.txt
gradio
torch
transformers
Pillow
peft
datasets
tqdm
